Add pool param to methods model interface
